datarmnet: msleep for 1ms, to avoid circular wait.
wda_svc_config and the netmgr qmi_client_setup,w hen run on
same CPU simultaneously, there is a chance that
wda_svc_arrive does not let the netmgr thread get the CPU.

Since netmgr thread is holding the rtnl_lock() and wda_svc_config
is waiting for the rtnl_lock(), this can lead to circular wait.

cond_resched() was used seems not reliable for all the possible cases.
when kworker on which wda_svc_config is an RT thread.

Change-Id: Ibf4d02bef9b3269f9972cfd214b2d693de9a21fa
Signed-off-by: Sharath Chandra Vurukala <quic_sharathv@quicinc.com>