msm: synx: optimize logic to set SYNX_ROOT
Current logic expects M as a relative path to kernel source and then
get the location of the source code path via $(KERNEL_SRC)/$(M).

But for LE, INPLACE_COMPILE will be set when calling build_module.sh
and M is set to absolute path.

Update logic to set source code path to makefile's path, so that it
can work in both LA and LE case.

Change-Id: Ic59373c782ebb21068375701129498add20efe9d
Signed-off-by: Qilong Wang <quic_qilowang@quicinc.com>